Rename integration to webserivce to match what it does
